backend/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
djoser
python-dotenv==0.19.0
drf-base64
Pillow==9.0.1
django-filter==21.1